requirements.txt
streamlit>=1.28.0
matplotlib>=3.7.0
seaborn>=0.12.0
urlextract>=1.8.0
wordcloud>=1.9.0
pandas>=2.0.0
emoji>=2.8.0
psycopg2-binary>=2.9.0
bcrypt>=4.0.0
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
requests>=2.31.0
reportlab>=4.0.0
python-docx>=0.8.11
plotly>=5.15.0
scikit-learn>=1.3.0
nltk>=3.8.1
textblob>=0.17.1
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
streamlit-option-menu>=0.3.6
streamlit-authenticator>=0.2.3
streamlit-aggrid>=0.3.4
streamlit-plotly-events>=0.0.6
